backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1

google-generativeai==0.8.3

langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-google-genai==2.0.5

chromadb==0.5.5

langchain-huggingface==0.1.1
sentence-transformers==2.6.1

pydantic==2.8.2
pydantic-core==2.20.1
typing_extensions==4.12.2

# PDF & OCR
PyPDF2==3.0.1
pymupdf==1.24.9
pillow==10.4.0
pytesseract==0.3.13
